Fixed warnings.

+++ src/Application/Controller/Controller.cpp
@@ -50,7 +50,7 @@ void Controller::Init()
     numControllers = (unsigned) configFile.GetListNum("controllers");
     for (nodeID = 0; nodeID < numControllers; nodeID++)
     {
-        str = configFile.GetListValue("controllers", nodeID, "");
+        str = configFile.GetListValue("controllers", (int) nodeID, "");
         endpoint.Set(str);
         CONTEXT_TRANSPORT->AddNode(nodeID, endpoint);
     }
